Remove custom `Payload` implementation in favor of circe `JsonObject`
The `Payload` implementation is almost identical to `io.circe.JsonObject`,
but the implementation and maintenance burden is quite high. The syntax
extensions also tend to clash in user code which is annoying.

The main motivation to have this in the first place, was to not depend on
circe in the core module. I no longer consider this worth it and rather
drop the custom Payload implementation and depend on circe instead.